potato_disease/Potato-Disease-Classification/api/requirements.txt
fastapi==0.116.1
uvicorn==0.23.2
numpy>=1.26,<1.27
pandas>=2.0,<2.1
tensorflow==2.20.0
keras==3.11.2
scikit-learn>=1.2,<1.3
opencv-python>=4.8,<4.9
Pillow>=10.0,<10.1
matplotlib>=3.7,<3.8
